+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Missing countries from 3.x: ARE, DNK, FIN, IND, ITA, SMR and SWE.
+
 ## [4.6.4] - 2020-10-06 [YANKED]
 
 ### Fixed