+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Address autocomplete for Argentina's capital.
+
 ## [3.15.0] - 2021-01-25
 
 ### Added